backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
pydantic-settings==2.8.1
python-jose[cryptography]==3.4.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
requests==2.32.3
python-dotenv==1.0.1
python-docx==1.1.2
pypdf==5.4.0
chromadb==0.6.3
pytest==8.3.5
pytest-asyncio==0.26.0
httpx==0.28.1
ruff==0.11.5
